Replace fairwins dashboard with POC components and merge mock data
